requirements.txt
Django==3.2.9
djangorestframework==3.12.4
django-redis==4.12.1
celery==5.1.2
redis==3.7.3
drf-yasg==1.20.0
flower==0.9.7